move SVM from ANALYSIS to MATH folder
