+++ src/openms/includes.cmake
@@ -7,6 +7,7 @@ include(source/SYSTEM/sources.cmake)
 include(source/MATH/STATISTICS/sources.cmake)
 include(source/MATH/MISC/sources.cmake)
 include(source/MATH/MISC/NNLS/sources.cmake)
+include(source/MATH/SVM/sources.cmake)
 include(source/DATASTRUCTURES/sources.cmake)
 include(source/METADATA/sources.cmake)
 include(source/METADATA/ID/sources.cmake)
@@ -23,7 +24,6 @@ include(source/FORMAT/sources.cmake)
 include(source/IONMOBILITY/sources.cmake)
 include(source/ANALYSIS/QUANTITATION/sources.cmake)
 include(source/ANALYSIS/SEQUENCE/sources.cmake)
-include(source/ANALYSIS/SVM/sources.cmake)
 include(source/ANALYSIS/MAPMATCHING/sources.cmake)
 include(source/ANALYSIS/DECHARGING/sources.cmake)
 include(source/ANALYSIS/ID/sources.cmake)
@@ -58,6 +58,7 @@ include(include/OpenMS/SYSTEM/sources.cmake)
 include(include/OpenMS/MATH/MISC/sources.cmake)
 include(include/OpenMS/MATH/MISC/NNLS/sources.cmake)
 include(include/OpenMS/MATH/STATISTICS/sources.cmake)
+include(include/OpenMS/MATH/SVM/sources.cmake)
 include(include/OpenMS/DATASTRUCTURES/sources.cmake)
 include(include/OpenMS/DATASTRUCTURES/Utils/sources.cmake)
 include(include/OpenMS/METADATA/sources.cmake)
@@ -79,7 +80,6 @@ include(include/OpenMS/ANALYSIS/ID/sources.cmake)
 include(include/OpenMS/ANALYSIS/MAPMATCHING/sources.cmake)
 include(include/OpenMS/ANALYSIS/QUANTITATION/sources.cmake)
 include(include/OpenMS/ANALYSIS/SEQUENCE/sources.cmake)
-include(include/OpenMS/ANALYSIS/SVM/sources.cmake)
 include(include/OpenMS/ANALYSIS/MRM/sources.cmake)
 include(include/OpenMS/ANALYSIS/TARGETED/sources.cmake)
 include(include/OpenMS/ANALYSIS/TOPDOWN/sources.cmake)
